calibrated motor speed min and max.

+++ Core/Common/Models/StructuresForRC.cpp
@@ -1,8 +1,8 @@
 struct payload
 {
     bool IsStopEngines;
-    int LeftJoystickX;
-    int LeftJoystickY;
-    int RightJoystickX;
-    int RightJoystickY;
+    uint8_t LeftJoystickX;
+    uint8_t LeftJoystickY;
+    uint8_t RightJoystickX;
+    uint8_t RightJoystickY;
 };
